requirements.txt
fastapi==0.61.1
grpcio==1.32.0
grpcio-tools==1.32.0
SQLAlchemy==1.3.19
python-dotenv==0.14.0
uvicorn==0.11.8
alembic==1.4.3
bcrypt==3.2.0
psycopg2-binary
